azalea_auth
Function auth

pub async fn auth(
    email: &str,
    opts: AuthOpts<'_>,
) -> Result<AuthResult, AuthError>

Authenticate with Microsoft. If the data isn’t cached, they’ll be asked to go to log into Microsoft in a web page.

The email is technically only used as a cache key, so it could be anything. You should just have it be the actual email so it’s not confusing though, and in case the Microsoft API does start providing the real email.

If you want to use your own code to cache or show the auth code to the user in a different way, use get_ms_link_code, get_ms_auth_token, get_minecraft_token and get_profile instead.