azalea_auth
Module sessionserver

Tell Mojang you’re joining a multiplayer server.

Structs

ForbiddenError

Enums

ClientSessionServerError

ServerSessionServerError

Functions

join

Tell Mojang’s servers that you are going to join a multiplayer server, which is required to join online-mode servers. The server ID is an empty string.

join_with_server_id_hash

serverside_auth

Ask Mojang’s servers if the player joining is authenticated. Included in the reply is the player’s skin and cape. The IP field is optional and equivalent to enabling ‘prevent-proxy-connections’ in server.properties